--- src/main/java/com/ormee/server/dto/quiz/SubmitRequestDto.java
@@ -1,12 +1,0 @@
-package com.ormee.server.dto.quiz;
-
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-public class SubmitRequestDto {
-    private String author;
-    private String password;
-    private List<SubmitDto> submissions;
-}
