+++ src/main/java/com/ormee/server/exception/ExceptionType.java
@@ -20,7 +20,6 @@ public enum ExceptionType {
     ASSIGNMENT_NOT_FOUND_EXCEPTION(HttpStatus.NOT_FOUND, "존재하지 않는 과제입니다."),
     FEEDBACK_NOT_FOUND_EXCEPTION(HttpStatus.NOT_FOUND, "존재하지 않는 피드백입니다."),
     STUDENT_LECTURE_NOT_FOUND_EXCEPTION(HttpStatus.NOT_FOUND, "존재하지 않는 학생-강의입니다."),
-    ASSIGNMENT_SUBMIT_NOT_FOUND_EXCEPTION(HttpStatus.NOT_FOUND, "존재하지 않는 과제 제출입니다."),
     PASSWORD_INVALID_EXCEPTION(HttpStatus.FORBIDDEN, "비밀번호가 일치하지 않습니다.");
 
     private final HttpStatus httpStatus;
