+++ src/main/java/com/ormee/server/repository/ProblemSubmitRepository.java
@@ -2,6 +2,7 @@ package com.ormee.server.repository;
 
 import com.ormee.server.model.Problem;
 import com.ormee.server.model.ProblemSubmit;
+import com.ormee.server.model.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,10 +11,10 @@ import java.util.Optional;
 
 @Repository
 public interface ProblemSubmitRepository extends JpaRepository<ProblemSubmit, Long> {
-    Optional<ProblemSubmit> findByProblemAndAuthorAndPassword(Problem problem, String author, String password);
     List<ProblemSubmit> findAllByProblem(Problem problem);
-    boolean existsByProblemAndAuthorAndPassword(Problem problem, String author, String password);
     Long countAllByProblemAndContentLike(Problem problem, String content);
 
     long countAllByProblem(Problem problem);
+
+    Optional<ProblemSubmit> findByProblemAndStudent(Problem problem, Student student);
 }
