+++ src/main/resources/application.properties
@@ -5,7 +5,7 @@ spring.datasource.username=postgres
 spring.datasource.password=ormee1826!
 spring.datasource.driver-class-name=org.postgresql.Driver
 
-spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=create
 
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.properties.hibernate.enable_lazy_load_no_trans=true
